Added initial example for Batty testing
This requires that you have Zorg installed to your system ahead
of time (use `pip install --upgrade .` if needed) and will run the
test API needed for Batty.

This does not yet include the `Loopback` connection and `Ping`
device that Batty is expecting. It does hack in the `echo` command
that the MCP should implement, but does not implement the `hello`
command on the robot.

https://github.com/hybridgroup/batty