Change request processing to use handler methods
Previously, the request processing used a beautiful block of `if`
statements that processed each level of the API. This has been
converted into some new logic for processing requests and the
variables within paths.

Right now, thanks to how the CPPP spec is written, the path of a
request is composed of resources (robots, commands, etc.) and the
identifiers for those (robot names, command names, etc.) that are
put together in pairs. This allows us to parse the request path
knowing that every odd section is a resource and every even section
is an identifier. With this, we determine the method name that
should be used for handling the request and pass in the identifiers
as arguments to the request.

-----

Fix JSON parsing of command arguments

Previously, command arguments were expected to be passed in as a
JSON object where the key was the argument name and the value was
the argument value. This worked well because Python has a solid
concept of named arguments, but it conflicted with the CPPP
specification.

Now command arguments are parsed using the Python `json` module
but it uses an `OrderedDict` to store the command arguments. The
command arguments are also now passed into the method as positional
arguments instead of named arguments, using the ordered values
provided in the request body.

This does not fully line up with the JSON specification, and a
ticket has been created in the CPPP repository at
https://github.com/hybridgroup/cppp-io/issues/5

-----

Add initial support for commands to the MCP

The Batty tests require that the MCP implements an `echo` command,
so initial commands have been added to the HTTP API. These are
tied to the HTTP API instead of the MCP within Zorg because it
doesn't actually exist.

This also adds support for commands in the MCP to the API, and
now passes the two tests provided by Batty for them.