Working on serializing connections
This adds the routes for accessing the robot's connections in the
API.

The tests are currently failing because we are not including
the `adaptor` key. This is because the Python adaptor name would not
match the JavaScript one, and it's not being included because the
CPPP specification does not currently require it. Including the
full path to the adaptor (the standard dotted path representation)
would also introduce the possibility of exposing too much of the
robot internals to the API.

We are also not including the `details` key, which appears to
reflect the options that are passed into the connection, because
the version of Batty that we are using does not test for it.
Eventually this will be added to the serialized output.