Added details to the API
The Batty tests are expecting an extra key to be passed through
when intializing connections and devices, so we are now passing
that through in the `details` key for both of these responses.

We are also now including the adaptor and driver name, but only
the last section of the dotted path. This was fairly easy to do
and just so happens to line up with what Batty is expecting.

Now, after installing the latest version of Batty from GitHub, we
are back to having a few passing tests and a bunch of failing ones.