Rework events within robots
**Breaking change:** `self.events` in drivers, robots, and the API
is now a dictionary mapping the event name (the key) to the events
settings (the value). The event settings are made up of a dictionary
containing the `source`, a Python function that is the source of
the events, and the `interval`, which is how often the event source
should be polled.

This adds support for events in the HTTP API which are handled
using the W3C EventSource specification [1] which allows for
server-sent events to be handled. Adding the support required
the creation of `EventStream` and `Event` classes.

The `EventStream` class allows for the creation of a separate
process which will poll the event source at a specific interval
and push the results (`Event` objects) into a queue to be read
later. `Event` objects hold the result of the event source and
allow them to be serialized to conform to the EventSource
specification and be read by clients. Right now the events
emitted through the `Event` classes are data-only events, so
they do not contain unique identifiers and do not have a specific
name.

There are plans to add later support for unique identifier and
named events which will be proposed for the CPPP specification.

This also creates an `EventsMixin` which can be used on any
object to allow it to hold events and create event sources. This
is currently used on the `http.Api`, `Driver`, and `Robot` classes
so their events can be handled in the API.

We are now down to two failing tests using Batty. These tests are
breaking because the MCP does not currently support the
`robot_created` and `robot_removed` events.

[1]: https://w3c.github.io/eventsource/